Made a hook for the isLegal function
